[REMOVE] Drop type-to-search because it got broken by a BC update with no easy fix
The `CommonKeyDown` was deprecated... with no suitable replacement.
The way `GameKeyDown` is done now is not easily hookable as it is directly used directly as the callback that is handed to DOM event handlers - before the `GameKeyDown` simply forwarded the call to the `CommonKeyDown`, so one could hook `CommonKeyDown` and it worked. Now one would need to go through all the DOM event handlers on all the objects that already exist and replace them manually.